Feat: Created components for food inventory codes for shortening purpose and also removed some unused lines.
